39: handle create new note by EditNotePresenter
