+++ app/src/main/kotlin/com/exallium/todoapp/allnotes/AllNotesViewImpl.kt
@@ -8,7 +8,7 @@ import butterknife.BindView
 import com.bluelinelabs.conductor.RouterTransaction
 import com.exallium.todoapp.R
 import com.exallium.todoapp.app.TodoApp
-import com.exallium.todoapp.createnote.CreateNoteViewImpl
+import com.exallium.todoapp.editnote.EditNoteViewImpl
 import com.exallium.todoapp.mvp.BaseViewImpl
 import com.exallium.todoapp.notedetail.NoteDetailViewImpl
 import com.jakewharton.rxbinding.view.clicks
@@ -44,7 +44,7 @@ class AllNotesViewImpl(bundle: Bundle) : BaseViewImpl<AllNotesView, AllNotesPres
     }
 
     override fun showCreateNote(args: Bundle) {
-        router.pushController(RouterTransaction.with(CreateNoteViewImpl(args)))
+        router.pushController(RouterTransaction.with(EditNoteViewImpl(args)))
     }
 
     override fun addNoteClicks(): Observable<Unit> = addNote.clicks()
