+++ app/src/main/kotlin/com/exallium/todoapp/editnote/EditNoteModel.kt
@@ -8,7 +8,7 @@ import rx.Single
  */
 interface EditNoteModel {
     fun getNote(id: String): Single<Note>
-    fun editNote(note: Note): Single<Unit>
+    fun saveNote(note: Note): Single<Unit>
 
     /**
      * creates in-memory Note object based on existing item and the new data
