+++ app/src/main/kotlin/com/exallium/todoapp/editnote/EditNotePresenter.kt
@@ -3,6 +3,7 @@ package com.exallium.todoapp.editnote
 import com.exallium.todoapp.R
 import com.exallium.todoapp.entities.Note
 import com.exallium.todoapp.mvp.BasePresenter
+import com.exallium.todoapp.repository.IdFactory
 import com.exallium.todoapp.screenbundle.ScreenBundleHelper
 import rx.Observable
 import rx.SingleSubscriber
@@ -13,24 +14,36 @@ import timber.log.Timber
  * Presenter for Edit Note Screen
  */
 class EditNotePresenter(view: EditNoteView,
-                          private val model: EditNoteModel,
-                          private val screenBundleHelper: ScreenBundleHelper) : BasePresenter<EditNoteView>(view) {
+                        private val model: EditNoteModel,
+                        private val screenBundleHelper: ScreenBundleHelper,
+                        private val idFactory: IdFactory) : BasePresenter<EditNoteView>(view) {
 
-    private val showNewNoteDetailSubscriberFn: (Unit?) -> (Unit) = {
-        view.showNewNoteDetail(getArgs())
+    private val showAllNotesSubscriberFn: (Unit?) -> (Unit) = {
+        view.showAllNotes(getArgs())
+    }
+
+    private val showNoteDetailSubscriberFn: (Unit?) -> (Unit) = {
+        view.showNoteDetail(getArgs())
     }
 
     override fun onViewCreated() {
-        screenBundleHelper.setTitle(getArgs(), R.string.edit_note_screen_title)
-        val noteId: String = screenBundleHelper.getNoteId(getArgs())
-
-        setupGetNoteDetailSubscription(noteId)
-
-        setupSaveNoteSubscription(noteId)
-
+        val args = getArgs()
+        if (screenBundleHelper.getNoteId(args) == null) {
+            screenBundleHelper.setTitle(args, R.string.create_note_screen_title)
+            setupSaveNewNoteSubscription()
+            setupCancelButtonClicksSubscription { showAllNotesSubscriberFn }
+        } else {
+            screenBundleHelper.setTitle(args, R.string.edit_note_screen_title)
+            val noteId: String = screenBundleHelper.getNoteId(args)!!
+            setupGetNoteDetailSubscription(noteId)
+            setupSaveEditedNoteSubscription(noteId)
+            setupCancelButtonClicksSubscription { showNoteDetailSubscriberFn }
+        }
         setupTextChangedSubscription()
+    }
 
-        view.cancelEditNoteClicks().map { null }.subscribe(showNewNoteDetailSubscriberFn).addToComposite()
+    fun setupCancelButtonClicksSubscription(cancelButtonFn: (Unit?) -> (Unit)) {
+        cancelButtonFn
     }
 
     fun setupTextChangedSubscription() {
@@ -42,6 +55,7 @@ class EditNotePresenter(view: EditNoteView,
                                    validationFn: (String) -> Boolean,
                                    showErrFn: () -> Unit) {
         textChangedObservable
+                .skip(1) // ignore initial load
                 .map(CharSequence::toString)
                 .map(validationFn)
                 .doOnNext { if (!it) {
@@ -66,7 +80,27 @@ class EditNotePresenter(view: EditNoteView,
         }).addToComposite()
     }
 
-    fun setupSaveNoteSubscription(noteId: String) {
+    fun setupSaveNewNoteSubscription() {
+        view.saveNoteClicks()
+                .flatMap { model.saveNote(buildNote()).toObservable() }
+                .subscribe(object : Subscriber<Unit>() {
+                    override fun onCompleted() {
+                        // do nothing
+                    }
+
+                    override fun onNext(unit: Unit) {
+                        Timber.d("saved new note")
+                        view.showNoteDetail(getArgs())
+                    }
+
+                    override fun onError(t: Throwable) {
+                        Timber.w(t, "error saving note")
+                        view.showUnableToSaveNoteError()
+                    }
+                }).addToComposite()
+    }
+
+    fun setupSaveEditedNoteSubscription(noteId: String) {
         model.getNote(noteId).subscribe(object : SingleSubscriber<Note>() {
             override fun onSuccess(note: Note) {
                 performSaveNote(note)
@@ -81,7 +115,7 @@ class EditNotePresenter(view: EditNoteView,
 
     private fun performSaveNote(oldNote: Note) {
         view.saveNoteClicks()
-                .flatMap { model.editNote(buildNote(oldNote)).toObservable() }
+                .flatMap { model.saveNote(buildNote(oldNote)).toObservable() }
                 .subscribe(object : Subscriber<Unit>() {
                     override fun onCompleted() {
                         // do nothing
@@ -89,7 +123,7 @@ class EditNotePresenter(view: EditNoteView,
 
                     override fun onNext(unit: Unit) {
                         Timber.d("saved note with id " + oldNote.id)
-                        view.showNewNoteDetail(getArgs())
+                        view.showNoteDetail(getArgs())
                     }
 
                     override fun onError(t: Throwable) {
@@ -99,9 +133,15 @@ class EditNotePresenter(view: EditNoteView,
                 }).addToComposite()
     }
 
-    internal fun buildNote(oldNote: Note): Note = model.buildNote(oldNote, view.getNewNoteTitle(), view.getNewNoteBody())
+    internal fun buildNote(oldNote: Note): Note = model.buildNote(oldNote, view.getNoteTitle(), view.getNoteBody())
+
+    internal fun buildNote(): Note {
+        val noteId = idFactory.createId().apply { screenBundleHelper.setNoteId(getArgs(), this) }
+        val now = System.currentTimeMillis()
+        return Note(noteId, view.getNoteTitle(), view.getNoteBody(), now, now)
+    }
 
     internal fun validateAllFields(): Boolean {
-        return model.validateNoteTitleText(view.getNewNoteTitle()) && model.validateNoteBodyText(view.getNewNoteBody())
+        return model.validateNoteTitleText(view.getNoteTitle()) && model.validateNoteBodyText(view.getNoteBody())
     }
 }
