+++ app/src/main/kotlin/com/exallium/todoapp/editnote/EditNoteViewImpl.kt
@@ -59,16 +59,16 @@ class EditNoteViewImpl(val bundle: Bundle) : BaseViewImpl<EditNoteView, EditNote
 
     override fun bodyTextChanges(): Observable<CharSequence> = noteBodyEditText.textChanges()
 
-    override fun showNewNoteDetail(args: Bundle) {
+    override fun showNoteDetail(args: Bundle) {
         view?.hideKeyboard(applicationContext?.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager)
         router.popCurrentController()
     }
 
-    override fun getNewNoteTitle(): String {
+    override fun getNoteTitle(): String {
         return noteTitleEditText.text.toString()
     }
 
-    override fun getNewNoteBody(): String {
+    override fun getNoteBody(): String {
         return noteBodyEditText.text.toString()
     }
 
@@ -90,5 +90,9 @@ class EditNoteViewImpl(val bundle: Bundle) : BaseViewImpl<EditNoteView, EditNote
 
     override fun toggleSubmit(enable: Boolean) {
         saveNote.isEnabled = enable
+    }
+
+    override fun showAllNotes(args: Bundle) {
+        router.popCurrentController()
     }
 }
