+++ app/src/main/kotlin/com/exallium/todoapp/screenbundle/ScreenBundleHelper.kt
@@ -11,5 +11,5 @@ interface ScreenBundleHelper {
     fun setTitle(bundle: Bundle, screenTitle: String)
     fun getTitle(bundle: Bundle?): String
     fun setNoteId(bundle: Bundle, noteId: String? = null)
-    fun getNoteId(bundle: Bundle): String
+    fun getNoteId(bundle: Bundle): String?
 }
