--- app/src/main/res/layout/create_note_view.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent"
-             android:orientation="vertical"
-             android:layout_margin="@dimen/note_detail_margin">
-
-    <EditText
-        android:id="@+id/create_note_title"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textStyle="bold"
-        android:hint="@string/create_note_title_hint" />
-
-    <EditText
-        android:id="@+id/create_note_body"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/note_detail_body_margin_top"
-        android:layout_marginBottom="@dimen/create_note_body_margin_bottom"
-        android:hint="@string/create_note_body_hint" />
-
-    <android.support.v7.widget.ButtonBarLayout
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:orientation="horizontal"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="@dimen/create_note_button_bar_layout_margin_top">
-
-        <Button
-            android:id="@+id/create_note_cancel_button"
-            style="?attr/buttonBarButtonStyle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="5"
-            android:text="@string/create_note_cancel_button_text" />
-
-        <android.support.v4.widget.Space
-            android:id="@+id/spacer"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:visibility="invisible"/>
-
-        <Button
-            android:id="@+id/create_note_save_button"
-            style="?attr/buttonBarButtonStyle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="5"
-            android:text="@string/create_note_save_button_text" />
-
-    </android.support.v7.widget.ButtonBarLayout>
-
-</LinearLayout>
