+++ app/src/main/res/values/strings.xml
@@ -13,17 +13,13 @@
     <string name="note_delete_button_text">Delete Note</string>
     <string name="note_edit_button_text">Edit Note</string>
 
-    <!-- Create Note -->
+    <!-- Create/Modify Note -->
     <string name="create_note_screen_title">New Note</string>
-    <string name="create_note_title_hint">title</string>
-    <string name="create_note_body_hint">body</string>
-    <string name="create_note_cancel_button_text">Cancel</string>
-    <string name="create_note_save_button_text">Save</string>
-
-    <!-- Edit Note -->
     <string name="edit_note_screen_title">Edit Note</string>
     <string name="invalid_note_title_error">Title cannot be empty</string>
     <string name="invalid_note_body_error">Body cannot be empty</string>
+    <string name="edit_note_title_hint">title</string>
+    <string name="edit_note_body_hint">body</string>
     <string name="edit_note_cancel_button_text">Cancel</string>
     <string name="edit_note_save_button_text">Save</string>
